pyproject.toml
[project]
name = "web-demo"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jealouscloud", email = "github@noaha.org" }
]
dependencies = [
    "flask>=3.1.0",
    "html-compose[full]>=0.6.2",
    "mistletoe>=1.4.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"web-demo" = "web:main"
"web-demo-live" = "web.server:live_reload"

[tool.rye.scripts]
"dev" = "python live-reload.py"
"prod" = "gunicorn backend.web.server:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy[mypyc]>=1.15.0",
    "rpython>=0.2.1",
    "watchfiles>=1.0.4",
    "gunicorn>=23.0.0",
]

[tool.hatch.metadata]
allow-direct-refereesnces = true

[tool.hatch.build.targets.wheel]
packages = ["backend/web"]
